feat: Add better event listeners
